+++ examples/basic/app/routes/demos/about/404.tsx
@@ -1,3 +1,0 @@
-export default function NotFound() {
-  return <div>404 Not Found</div>;
-}