+++ examples/basic/app/routes/demos/about/index.tsx
@@ -1,3 +1,7 @@
 export default function DemosAboutIndex() {
   return <div>/demos/about/index.jsx</div>;
 }
+
+export function ErrorBoundary() {
+  return <div> Oops!</div>;
+}